Flappy bird!!
This is game developed using awt package and applets in java. This project uses the concept of multithreading in java. In this game we have to press space bar to move the ball in the upward direction. Their are two classes in this project compile both the classes using "javac Ball.java" and "javac Checkiii.java" . Now when you get class file run the applet using command "appletviewer Checkiii.java'.